backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
django-filter==22.1
djoser==2.1.0
python-dotenv==0.21.0
reportlab==3.6.11
pytest==7.1.3
pytest-django==4.5.2
gunicorn==20.1.0
psycopg2-binary==2.9.4
django-colorfield==0.7.2